uiwebapp/requirements.txt
mysql-connector-python
flask
Flask-cors
pandas
requests
